+++ src/app/demo/ellipsis/index.ts
@@ -1,0 +1,2 @@
+export * from './ellipsis-demo.component';
+export * from './ellipsis-demo.module';