+++ src/app/exports/common/index.ts
@@ -1,5 +1,6 @@
 export * from './auto-focus.directive';
 export * from './dropdown.directive';
+export * from './date.pipe';
 export * from './router-reuse-strategy';
 export * from './assets-loader.service';
 export * from './resizeable.directive';