+++ src/app/exports/ellipsis/index.ts
@@ -1,0 +1,2 @@
+export * from './ellipsis.component';
+export * from './ellipsis.module';