+++ src/app/exports/rebirth-ng.config.ts
@@ -61,8 +61,15 @@ export class RebirthNGConfig {
   dialog = {
     button: {
       yes: 'Yes',
-      no: 'No'
+      btnYesType: 'primary',
+      no: 'No',
+      btnNoType: 'warning'
     }
+  };
+
+  ellipsis = {
+    length: 60,
+    placement: 'top'
   };
 
   fileUpload = {
@@ -74,6 +81,10 @@ export class RebirthNGConfig {
     chooseButton: 'Choose',
     uploadButton: 'Upload',
     cancelButton: 'Cancel'
+  };
+
+  modal = {
+    animation: true
   };
 
   pager = {